LayoutTests/canvas/philip/tests/2d.line.cap.valid-expected.txt

< [index] >
2d.line.cap.valid
Setting lineCap to valid values works
References: 2d.lineCap.set 2d.lineCap.get
Actual output:
Passed
